Take the first coefficient norm in double precision
The magnitude of the first DFT coefficient, which the translation stage uses as
its convergence criterion, was computed with sqrtf. Its argument and its
destination are both double, so the only effect was to discard precision: the
value came back 3.49e-08 out relative.

Fixes TranslationStage.FirstCoefficientNormIsDoublePrecision.

+++ fsm_lo/include/fsm_core.hpp
@@ -2673,7 +2673,7 @@ class Translation
     /* X1 */
     std::vector<double> X1 = DFTUtils::getFirstDFTCoefficient(diff, r2rp);
 
-    *norm_x1 = sqrtf(X1[0]*X1[0] + X1[1]*X1[1]);
+    *norm_x1 = sqrt(X1[0]*X1[0] + X1[1]*X1[1]);
 
     /* Find the x-wise and y-wise errors */
     double t = M_PI + current_t;
